backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
chromadb==0.4.22
pymongo==4.6.1
ollama==0.1.6
lxml==5.1.0
sentence-transformers
